requirements.txt
enum34~=1.0.4
cffi>=1.4.1,<2.0.0
pyasn1~=0.1.9
six~=1.10.0
pycparser~=2.14.0
backports.ssl-match-hostname~=3.4.0.2
idna~=2.0.0
cryptography~=1.5.0
certifi~=2016.2.28
dj-static~=0.0.6
Django~=1.8.11
django-custom-user~=0.5.0
django-extensions~=1.5.7
django-filter~=0.15.3
django-jinja~=1.4.1
django-macaddress~=1.3.2
django-rest-swagger~=2.1.2
django-smart-selects~=1.3.0
djangorestframework~=3.5.0
djangorestframework-bulk~=0.2.1
drf-extensions~=0.2.8
drf-nested-routers~=0.11.1
gevent~=1.0.2
gunicorn~=19.3.0
greenlet~=0.4.9
ipaddress~=1.0.14
ipython~=3.1.0
Jinja2~=2.8.0
logan~=0.7.1
MarkupSafe~=0.23.0
netaddr~=0.7.18
PyYAML~=3.11.0
static3~=0.6.1
coreapi~=2.3.1
coreschema~=0.0.4
itypes~=1.1.0
uritemplate~=3.0.0
openapi-codec~=1.3.2
simplejson~=3.11.1
